requirements.txt
Django
psycopg2-binary
django-leaflet
python-dotenv
gunicorn
dj-database-url
whitenoise
